+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## New features
 
+* Add support for EnergyPlus v25.1.0.
+
 * Add support for EnergyPlus v24.2.0 (#603).
 
 * Add support for EnergyPlus v24.1.0 (#582, #602).